+++ src/task3.cpp
@@ -23,7 +23,7 @@ bool checkPrime(unsigned int value)  //проверка числа на простоту
 */
 
 
-unsigned long sumPrime(unsigned int hbound)  //сумма всех простых чисел до hbound (не включая его)
+unsigned long long sumPrime(unsigned int hbound)  //сумма всех простых чисел до hbound (не включая его)
 {
 	unsigned long sumPrimeNumbers = 0;
 	for (int i = 2; i < hbound; i++)
